backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
sentence-transformers
vaderSentiment
numpy
scikit-learn
python-dotenv
